chore: Initial setup
We are making bold decisions with this Pull Request. By picking Bottle as
the framework, Pipfile (Pipenv) as our dependency tool, and "enforcing"
the use of pre-commit and Conventional Commits.

It might sound pretentious (and it is), but it's a good opportunity to
exercise those practices used by famous companies.

connects to #1