feat: Add a MVP version of the timer
With the help of Javascript and jQuery, we now have a first version of
the timer working fine. This should be faced as a Proof of Concept. It
still needs tests, the transpiling process to a "backwards
compatibility" Javascript, and some sort of dependency management via
NPM or yarn.

connects to #3