refactor: Remove jQuery dependency (#15)
It was going to be a PR about bringing a functional approach to the Javascript code. But after some
attempts, it's not that easy to apply immutability and event-driven without some sort of state
management (like Redux or MobX). I don't think it's the right time to bring this complexity to the
project, so this PR is just about removing jQuery and organizing the Timer class a little bit.