feat: Authenticate the user
This is an enabler. It doesn't affect the user experience we have so
far, but it'll allow us to progress with other features (like reports
and custom properties).

By using Github OAuth system, we are now able to identify the user, and
keep his/her email in a server session.

There are two major things that we need to address asap:
- Add a new layer of tests that validates the communication between our
client and the auth server (integration tests)
- Better logging and control of the AuthException error

closes #20