feat(auth): Store the user on the database
In order to show the user's progress we need first to store her/him. In this PR we implement that,
by letting the transactional control to the view (instead of adding it as the Model's
responsibility). It would be great to have a different layer of abstraction to deal with this kind
of thing (instead of relying on staticmethods/classmethods).

connects to #21