fix(auth): Use login field instead of email while connecting to Github
The email field is an optional field on Github. We've been using this field as some sort of "foreign
key" to connect "our" user with the Github's account. With this change, we are now pointing to the
login/username field, which is (afaik) a required field on Github API.

fix #65