chore: Use a cookiecutter like project template folder
With the PWA task, I'm facing a dilema about the statics organization and how are going to deal with
some stuff served in the root path of the URL (e.g. /service-workers.js). This PR proposes an
official standard for our folders, which can reduce confusion when facing views serving files
instead of them being created inside the static folder.

connects to #27